+++ CustomCritters/docs/release-notes.md
@@ -1,6 +1,9 @@
 ﻿﻿[← back to readme](README.md)
 
 # Release notes
+## Upcoming release
+* Added integration with BugNet so you can catch custom critters.
+
 ## 1.2.8
 Released 19 June 2021 for Stardew Valley 1.5. Updated by Pathoschild.